Consolidate cli base classes
