+++ apps/crc_idle.py
@@ -8,7 +8,7 @@ Resource summaries are provided for GPU and CPU partitions.
 from argparse import Namespace
 from typing import Dict, Tuple
 
-from ._base_parser import BaseParser
+from .utils.cli import BaseParser
 from .utils.system_info import Shell, Slurm
