+++ apps/crc_quota.py
@@ -18,7 +18,7 @@ import sys
 from argparse import Namespace
 from typing import Optional, Tuple
 
-from ._base_parser import BaseParser
+from .utils.cli import BaseParser
 from .utils.system_info import Shell
