+++ apps/crc_show_config.py
@@ -3,7 +3,7 @@
 from argparse import Namespace
 from typing import Dict
 
-from ._base_parser import BaseParser
+from .utils.cli import BaseParser
 from .utils.system_info import Shell, Slurm
