+++ apps/crc_sus.py
@@ -9,7 +9,7 @@ from typing import Dict
 
 import dataset
 
-from ._base_parser import BaseParser
+from .utils.cli import BaseParser
 from .utils.system_info import Shell, Slurm
