+++ apps/utils/cli.py
@@ -1,26 +1,49 @@
-"""Utility classes for building commandline parsers/applications."""
+"""Parent classes used when building command line applications."""
 
 import abc
+import os
 import sys
 from argparse import ArgumentParser, Namespace, HelpFormatter
+from textwrap import dedent
+from typing import Optional, List
 
 from .. import __version__
 
 
-class Parser(ArgumentParser):
-    """Extends the builtin ``ArgumentParser`` class with better formatting and error handling
+class BaseParser(ArgumentParser):
+    """Base class for building command line applications.
 
-    Instance are automatically created with a ``--version`` argument.
+    Inheriting from this class ensures child applications behave consistently
+    and share the same version number.
     """
 
-    def __init__(self, *args, **kwargs) -> None:
-        """Define default arguments for the commandline interface"""
+    def __init__(self) -> None:
+        """Define arguments for the command line interface"""
 
         # Increase the default width of the application help text
         formatter_factory = lambda prog: HelpFormatter(prog, max_help_position=50)
+        super(BaseParser, self).__init__(formatter_class=formatter_factory)
 
-        super().__init__(*args, **kwargs, formatter_class=formatter_factory)
-        self.add_argument('--version', action='version', version=f'{self.prog} version {__version__}')
+        # Strip indent from class docs and use as application description
+        self.description = '\n'.join(dedent(paragraph) for paragraph in self.__doc__.split('\n'))
+        self.add_argument('-v', '--version', action='version', version=self.app_version)
+
+    @property
+    def app_version(self) -> str:
+        """Return the application name and version as a string"""
+
+        program_name = os.path.splitext(self.prog)[0]
+        return '{} version {}'.format(program_name, __version__)
+
+    @abc.abstractmethod
+    def app_logic(self, args: Namespace) -> None:
+        """Logic to evaluate when executing the application
+
+        Args:
+            args: Parsed command line arguments
+        """
+
+        raise NotImplementedError
 
     def error(self, message: str) -> None:
         """Handle parsing errors and exit the application
@@ -45,43 +68,16 @@ class Parser(ArgumentParser):
 
         raise SystemExit(message)
 
-
-class CommandlineApplication(metaclass=abc.ABCMeta):
-    """Base class for building command line applications.
-
-    Subclasses should define the following:
-
-    1. The commandline interface via the ``app_interface`` method
-    2. The top level application logic via the ``app_logic`` method
-    """
-
-    def __init__(self):
-        """Initialize the commandline application"""
-
-        self.parser = Parser(description=self.__doc__.split('\n')[0])
-        self.app_interface(self.parser)
-
-    def app_interface(self, parser: Parser) -> None:
-        """Define the application commandline interface
-
-        Args:
-            parser: Argument parser used by the parent commandline application
-        """
-
-    @abc.abstractmethod
-    def app_logic(self, args: Namespace) -> None:
-        """Logic to evaluate when executing the application
-
-        Args:
-            args: Parsed command line arguments
-        """
-
     @classmethod
-    def execute(cls) -> None:
-        """Parse command line arguments and execute the application"""
+    def execute(cls, args: Optional[List[str]] = None) -> None:
+        """Parse command line arguments and execute the application
+
+        Args:
+            args: Optionally parse the given arguments instead of reading STDIN
+        """
 
         app = cls()
-        args = app.parser.parse_args()
+        args = app.parse_args(args)
 
         try:
             app.app_logic(args)
