+++ tests/test_crc_usage.py
@@ -1,9 +1,8 @@
 """Tests for the ``CrcUsage`` class."""
 
+import grp
 import os
 from unittest import TestCase, skipIf
-
-import grp
 
 from apps.crc_usage import CrcUsage
 from apps.utils.system_info import Slurm
@@ -15,16 +14,14 @@ class AccountArgument(TestCase):
     def test_default_account(self) -> None:
         """Test the default account matches the current user's primary group"""
 
-        app_parser = CrcUsage().parser
-        parsed_account = app_parser.parse_args([]).account
+        parsed_account = CrcUsage().parse_args([]).account
         current_account = grp.getgrgid(os.getgid()).gr_name
         self.assertEqual(current_account, parsed_account)
 
     def test_custom_account_name(self) -> None:
         """Test a custom account is used when specified"""
 
-        app_parser = CrcUsage().parser
-        parsed_account = app_parser.parse_args(['dummy_account']).account
+        parsed_account = CrcUsage().parse_args(['dummy_account']).account
         self.assertEqual('dummy_account', parsed_account)
 
 
@@ -36,7 +33,7 @@ class MissingAccountError(TestCase):
         """Test a ``RuntimeError`` is raised for a missing slurm account"""
 
         app = CrcUsage()
-        args = app.parser.parse_args(['dummy_account'])
+        args = app.parse_args(['dummy_account'])
 
         with self.assertRaisesRegex(RuntimeError, 'No slurm account was found'):
             app.app_logic(args)
